atol

+++ 0x0C-more_malloc_free/101-mul.c
@@ -31,7 +31,7 @@ exit(98);
 }
 }
 }
-mul = atoi(argv[1]) * atoi(argv[2]);
+mul = atol(argv[1]) * atol(argv[2]);
 printf("%ld\n", mul);
 }
 return (0);
